Fikser visning av preutfylt dato fra eksisterende korrigert vedtak
